+++ src/configuration/index.ts
@@ -1,1 +1,2 @@
+export * from './option_splitter'
 export * from './types'
