+++ src/configuration/option_splitter_spec.ts
@@ -1,6 +1,6 @@
 import { describe, it } from 'mocha'
 import { expect } from 'chai'
-import OptionSplitter from './option_splitter'
+import { OptionSplitter } from './option_splitter'
 
 describe('OptionSplitter', () => {
   const examples = [
